+++ src/config/keybinds.rs
@@ -331,6 +331,8 @@ pub struct Keybinds {
     pub switch_tab: Vec<IndexedKeybind>,
     pub switch_workspace: Vec<IndexedKeybind>,
     pub close_tab: ActionKeybinds,
+    pub move_tab_left: ActionKeybinds,
+    pub move_tab_right: ActionKeybinds,
     pub rename_pane: ActionKeybinds,
     pub edit_scrollback: ActionKeybinds,
     pub copy_mode: ActionKeybinds,
@@ -493,6 +495,8 @@ impl Config {
             switch_tab: Vec::new(),
             switch_workspace: Vec::new(),
             close_tab: empty_action!(),
+            move_tab_left: empty_action!(),
+            move_tab_right: empty_action!(),
             rename_pane: empty_action!(),
             edit_scrollback: empty_action!(),
             copy_mode: empty_action!(),
@@ -634,6 +638,8 @@ impl Config {
                 source
             );
             apply_action!(keybinds.close_tab, close_tab, source);
+            apply_action!(keybinds.move_tab_left, move_tab_left, source);
+            apply_action!(keybinds.move_tab_right, move_tab_right, source);
             apply_action!(keybinds.rename_pane, rename_pane, source);
             apply_action!(keybinds.edit_scrollback, edit_scrollback, source);
             apply_action!(keybinds.copy_mode, copy_mode, source);